Create hash.cpp
Hash Table
https://www.youtube.com/watch?v=KyUTuwz_b7Q
Data structure – fast retrieving of data regardless of how big the data set is.
Used widely.
Sequential search use brute-force algorithm, checking every element from the beginning. It would take a long
time if a data set is large.
Time to look up is independent of the size of the array.
How to populate a hash table?
Calculate an index from the key
Assume the table size is 11.
If key is name,
Mia 77 + 105 + 97 = 279 279%11 = 4 à Put Mia at index 4.
If key is a number
123 123%11 = 2 à Put 123 at index 2.
If a telephone number
760-740-1234 760+740+1234%11
Depending on key, you need to come up with a good way to calculate an index.
How to look up a key?
Mia 77 + 105 + 97 = 279 279%11 = 4 à Mia can be found at 4.
Collision
Multiple keys could map to the same index.
How to deal with collisions?
Open addressing – linear probing, put it in the next empty slot. If it reaches the end of the array, wrap
around to index 0.
Plus 3 rehash – rather than checking the next slot, check every 3rd slot.
Quadratic robing – (failed attempt)^2 – each time another look up causes a collision, the next
look up is farther than the original index. i.e. index = 3, 4, 7, 12, …
Double hashing – if collide, apply another hashing function (key%differentConstant) and add
this to the original index.
How to find key if open addressing is used to build the table?
Calculate the index from the key. Go to that index. If it is not found, use sequential search
starting at the next slot until it finds the key.
How to reduce collisions?
Make a table bigger than the number of items. For example, 75% occupied.
Number of items/size of array = load factor.
If the table is a dynamic structure, it could be made to resize automatically when a load factor
reaches a threshold.
Minimize collisions. Pick a good hush function to calculate an index.
Hash function should be easy to calculate.
Closed addressing (a.k.a. chaining) – each slot in the table is a linked list. Items with the same index
form a linked list.
How to find key if closed addressing is used?
Calculate the index from the key. Go to that index. Traverse the linked list until key is found.
No more collisions but clustering may be an issue. – a bunch of items belonging to the same
index and there are many empty slots in the table.
Uniform distribution. Pick a good hush function to calculate an index.
Which addressing to use?
Traversing a linked list may be time consuming.
If a load factor is low, open addressing might be better.
Time complexity of searching by hash table
Best-case – every key is at the calculated index. O(1)
Worst case – all items maps to the same index. O(n)
Time complexity of insert and delete
Best-case – every key is at the calculated index. O(1)
Worst case – all items maps to the same index. O(n)
Space complexity
O(n)